% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{addContradictionCount}
\alias{addContradictionCount}
\title{addContradictionCount}
\usage{
addContradictionCount(metrics)
}
\description{
Function to count and add detected contradictions as well as performed DQ checks.
}
