% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{addStatistic}
\alias{addStatistic}
\title{addStatistic}
\usage{
addStatistic(bdata, col, row)
}
\description{
Funtion to calculate the overall DQ metrics for the completeness and plausibility dimensions
}
