% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{caseCompletenessIndicator}
\alias{caseCompletenessIndicator}
\title{caseCompletenessIndicator}
\usage{
caseCompletenessIndicator(vm_case, vm_case_misg)
}
\description{
This function calculates the indicator for Case Completeness (dqi_cc_ccr), and adds related metadata and parameters.
}
