% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{checkAgePlausibility}
\alias{checkAgePlausibility}
\title{checkAgePlausibility}
\usage{
checkAgePlausibility(birthDate, currentData, ageMax)
}
\description{
This function checks the data values related to the data item "birthdate" for implausible values and returns the detected plausibility issues.
}
