% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCVD.R
\name{checkContradictionRules}
\alias{checkContradictionRules}
\title{checkContradictionRules}
\usage{
checkContradictionRules(df, mathRules, logicRules, contraCol)
}
\description{
This function checks the semantic plausibly of loaded data set using predefined DQ rules.
}
