% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordDqLib.R
\name{checkCordDQ}
\alias{checkCordDQ}
\title{checkCordDQ}
\usage{
checkCordDQ(
  instID,
  reportYear,
  inpatientCases,
  refData1,
  refData2,
  dqInd,
  repCol,
  cl,
  bItemCl,
  totalRow,
  oItem,
  ...
)
}
\description{
This function checks the quality of loaded data regarding selected quality metrics
The default data quality dimensions are completeness, plausibility, uniqueness and concordance
}
