% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{checkDifferenceRule}
\alias{checkDifferenceRule}
\title{checkDifferenceRule}
\usage{
checkDifferenceRule(studyData, item1, item2, min, max, ...)
}
\description{
Function to detect contradictions using a predefined difference rule.
}
