% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{checkFutureDate}
\alias{checkFutureDate}
\title{checkFutureDate}
\usage{
checkFutureDate(dateValues)
}
\description{
This function checks the loaded temporal data for implausible values and returns the values dated to the future.
}
