% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{checkLogicalRule}
\alias{checkLogicalRule}
\title{checkLogicalRule}
\usage{
checkLogicalRule(rID, contra, contraCol, item1, value1, item2, value2, ...)
}
\description{
Function to detect contradictions using a predefined logical rule.
}
