% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{checkMathRule}
\alias{checkMathRule}
\title{checkMathRule}
\usage{
checkMathRule(rID, contra, contraCol, item1, operator, item2, min, max, ...)
}
\description{
Function to detect contradictions using a predefined mathematical rule.
}
