% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordDqLib.R
\name{checkOrphaCodingCompleteness}
\alias{checkOrphaCodingCompleteness}
\title{checkOrphaCodingCompleteness}
\usage{
checkOrphaCodingCompleteness(refData, cl)
}
\description{
This function checks the completeness of OrphaCoding
}
