% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{checkOrphaCodingPlausibility}
\alias{checkOrphaCodingPlausibility}
\title{checkOrphaCodingPlausibility}
\usage{
checkOrphaCodingPlausibility(refData2, cl)
}
\description{
This function checks the plausibility of ICD-Orpha links.
}
