% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{checkRangeRule}
\alias{checkRangeRule}
\title{checkRangeRule}
\usage{
checkRangeRule(ndata, rID, itemCol, outlierCol, item, min, max, ...)
}
\description{
Function to detect outliers using a predefined range rule.
}
