% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCVD.R
\name{checkRangeRules}
\alias{checkRangeRules}
\title{checkMissingRules}
\usage{
checkRangeRules(ndata, rRules, imCol, outlierCol)
}
\description{
This function checks the range plausibly using predefined DQ rules.
}
