% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{checkSemanticUniqueness}
\alias{checkSemanticUniqueness}
\title{checkSemanticUniqueness}
\usage{
checkSemanticUniqueness(refData1, refData2, cl)
}
\description{
This function checks the quality of loaded data regarding the semantic uniqueness.
}
