% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{checkSubjCompleteness}
\alias{checkSubjCompleteness}
\title{checkSubjCompleteness}
\usage{
checkSubjCompleteness(subj, itemVec)
}
\description{
Function to check the completeness of recorded subjects such as inpatient or outpatients.
}
