% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{checkSymbolicConjunctions}
\alias{checkSymbolicConjunctions}
\title{checkSymbolicConjunctions}
\usage{
checkSymbolicConjunctions(item1, value1, item2, value2, ...)
}
\description{
Function to detect contradictions using predefined symbolic conjunctions.
}
