% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{checkTemporalPlausibility}
\alias{checkTemporalPlausibility}
\title{checkTemporalPlausibility}
\usage{
checkTemporalPlausibility(dataset, itemCol, cl)
}
\description{
This function checks the plausibility of date values in a given data set.
}
