% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCord.R
\name{checkUniqueIcdOrphaCoding}
\alias{checkUniqueIcdOrphaCoding}
\title{checkUniqueIcdOrphaCoding}
\usage{
checkUniqueIcdOrphaCoding(refData1, refData2, cl)
}
\description{
This function checks the uniqueness of RD cases coded with ICD-Orpha links
}
