% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCVD.R
\name{cvdDqChecker}
\alias{cvdDqChecker}
\title{cvdDqChecker}
\usage{
cvdDqChecker(rPath, ruleMeta)
}
\description{
This function checks the quality of loaded data set regarding common DQ issues that may arise in the context of common diseases like cardiovascular diseases (CVDs).
}
