% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{deprecatedMetrics}
\alias{deprecatedMetrics}
\title{deprecatedMetrics}
\usage{
deprecatedMetrics(df)
}
\description{
Function to handle deprecated DQ metrics that were renamed according to the abbreviations specified in the publication DOI:10.1055/a-2006-1018.
The aim of this function is to support legacy versions and migrate from deprecated code without introducing bugs.
}
