% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{dqChecker}
\alias{dqChecker}
\title{dqChecker}
\usage{
dqChecker(
  studyData,
  coreDomain,
  numItems,
  catItems,
  tempItems,
  itemMandatory,
  itemOptional,
  missingCode,
  ...
)
}
\description{
This function assesses the quality of loaded data using generic indicators and domain-specific DQ checks.
The DQ indicators were implemented based on the DQ concept published under the DOI:10.1055/a-2006-1018.
}
