% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{getAgeMaxOutlierx}
\alias{getAgeMaxOutlierx}
\title{getAgeMaxOutlier}
\usage{
getAgeMaxOutlierx(dItem1.vec, dItem2.vec, n)
}
\description{
This Function checks the data item "birthdate" for implausible values
}
