% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCord.R
\name{getCaseCompletenessRate}
\alias{getCaseCompletenessRate}
\title{getCaseCompletenessRate}
\usage{
getCaseCompletenessRate(cdata, ddata, caseItems)
}
\description{
This function evaluates the completeness of case module
}
