% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCord.R
\name{getConcIndicator}
\alias{getConcIndicator}
\title{getConcIndicator}
\usage{
getConcIndicator(dist, index)
}
\description{
This function calculates the z-score value to measure concordance indicators such as the concordance of RD cases or the concordance of tracer cases
}
