% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{getConcWithRefValues}
\alias{getConcWithRefValues}
\title{getConcWithRefValues}
\usage{
getConcWithRefValues(tracerCase_rel, concRef)
}
\description{
This function evaluates the concordance of tracer cases with reference values from the literature of national references.
}
