% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCVD.R
\name{getContradictionRules}
\alias{getContradictionRules}
\title{getContradictionRules}
\usage{
getContradictionRules(path, sheetName)
}
\description{
Functions to extract contradiction rules from spreadsheets.
}
