% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordDqLib.R
\name{getExtendedReport}
\alias{getExtendedReport}
\title{getExtendedReport}
\usage{
getExtendedReport(repCol, cl, td, useCase, path)
}
\description{
This function generates an extended data quality reports with infos about Projecathon use cases
}
