% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{getLongReport}
\alias{getLongReport}
\title{getLongReport}
\usage{
getLongReport(metrics, sheetList, design, path)
}
\description{
This function generates detailed reports on user-selected metrics and DQ issues including outliers, contradictions, missing values, and missing items.
}
