% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{getMissingItem}
\alias{getMissingItem}
\title{getMissingItem}
\usage{
getMissingItem(mandatoryItems)
}
\description{
This function checks the loaded data items for completeness issues and returns the detected missing data items.
}
