% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{getMissingValue}
\alias{getMissingValue}
\title{getMissingValue}
\usage{
getMissingValue(df, bItemCol, outCol1, outCol2, ...)
}
\description{
This function checks the loaded data for missing data values. It also supports missing rules.
}
