% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{getPatRecordMetrics}
\alias{getPatRecordMetrics}
\title{#' @title getPatRecordMetrics}
\usage{
getPatRecordMetrics(identifier)
}
\description{
This function provides the number of analyzed patients and the overall number of patients with DQ issues.
}
