% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordDqLib.R
\name{getReport}
\alias{getReport}
\title{geReport}
\usage{
getReport(repCol, cl, td, path)
}
\description{
This function generates data quality reports about detected quality issues, user selected indicators and key numbers
}
