% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCord.R
\name{getSubjCompleteness}
\alias{getSubjCompleteness}
\title{getSubjCompleteness}
\usage{
getSubjCompleteness(rep, subj, itemVec, medData)
}
\description{
This function evaluates the completeness of recorded subjects such as inpatient or outpatients
}
