% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{getTotalStatistic}
\alias{getTotalStatistic}
\title{getTotalStatistic}
\usage{
getTotalStatistic(col, row)
}
\description{
Function to aggregate the different types of analyzed data and calculate the overall DQ metrics.
}
