% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{getUserSelectedMetrics}
\alias{getUserSelectedMetrics}
\title{getUserSelectedMetrics}
\usage{
getUserSelectedMetrics(dqInd, tdata)
}
\description{
This function enable users to select desired DQ metrics
}
