% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{itemCompletenessIndicator}
\alias{itemCompletenessIndicator}
\title{itemCompletenessIndicator}
\usage{
itemCompletenessIndicator(im, im_misg)
}
\description{
This function calculates the Item Completeness Rate (dqi_cc_icr), a generic indicator that assesses the completeness of mandatory data items, and adds related metadata and DQ parameters.
}
