% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{missingCheck}
\alias{missingCheck}
\title{missingCheck}
\usage{
missingCheck(df, item, itemCol, cl1, cl2, ...)
}
\description{
Function to check individual data items for missing values. The new version also supports coded missing values and data blanking rules. Data values hidden due to data blanking rules are not considered missing values.
}
