% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{orphaCompletenessIndicator}
\alias{orphaCompletenessIndicator}
\title{orphaCompletenessIndicator}
\usage{
orphaCompletenessIndicator(icd_tracer, oc_misg)
}
\description{
This function calculates the Orphacoding Completeness Rate  (dqi_cc_ocr), a specific indicator for RD coding, and adds related metadata and DQ parameters.
}
