% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{orphaPlausibilityIndicator}
\alias{orphaPlausibilityIndicator}
\title{orphaPlausibilityIndicator}
\usage{
orphaPlausibilityIndicator(link, link_ip)
}
\description{
This function calculates the Orphacoding Plausibility Rate (dqi_pl_opr), a specific indicator for assessing the semantic plausibility of RD diagnoses, and adds related metadata and DQ parameters.
}
