% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{rangePlausibilityIndicator}
\alias{rangePlausibilityIndicator}
\title{rangePlausibilityIndicator}
\usage{
rangePlausibilityIndicator(v_slc, v_ip)
}
\description{
This function calculates the Range Plausibility Indicator (dqi_pl_rpr), a generic metric to assess the range plausibility of selected data values. It also adds related metadata and parameters in the DQ report.
}
