% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{rdCaseDissimilarityIndicator}
\alias{rdCaseDissimilarityIndicator}
\title{rdCaseDissimilarityIndicator}
\usage{
rdCaseDissimilarityIndicator(rdCase, rdCase_dup)
}
\description{
This function calculates the RD Case Dissimilarity Rate(dqi_un_cdr), a specific indicator for assessing the syntactic uniqueness of RD diagnoses, and provides related metadata and DQ parameters.
}
