% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{rdCaseUnambiguityIndicator}
\alias{rdCaseUnambiguityIndicator}
\title{rdCaseUnambiguityIndicator}
\usage{
rdCaseUnambiguityIndicator(rdCase, rdCase_amb)
}
\description{
This function calculates the RD Case unambiguity Rate (dqi_un_cur), a specific indicator for assessing the semantic uniqueness of RD diagnoses, and adds related metadata and DQ parameters.
}
