% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{rdConcordanceMetrics}
\alias{rdConcordanceMetrics}
\title{rdConcordanceMetrics}
\usage{
rdConcordanceMetrics(year, ipat, tracerCase, rdCase, orphaVec, cl)
}
\description{
This function provides normalized RD-specific metrics for the concordance dimension (D4).
}
