% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{rdDqChecker}
\alias{rdDqChecker}
\title{rdDqChecker}
\usage{
rdDqChecker(tracerRef, rdStandard)
}
\description{
This function checks the quality of loaded data regarding DQ issues that may arise in context of rare diseases (RDs).
The default DQ dimensions are completeness and plausibility.
}
