% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{semanticPlausibilityIndicator}
\alias{semanticPlausibilityIndicator}
\title{semanticPlausibilityIndicator}
\usage{
semanticPlausibilityIndicator(vs_cd, vc)
}
\description{
This function calculates the Semantic Plausibility Rate (dqi_pl_spr), a generic indicator to assess the semantic plausibility of selected data values, and adds related metadata and DQ parameters.
}
