% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{subjectCompletenessIndicator}
\alias{subjectCompletenessIndicator}
\title{subjectCompletenessIndicator}
\usage{
subjectCompletenessIndicator(s, s_inc)
}
\description{
Function to calculate the indicator for Subject Completeness (dqi_cc_scr), and adds related metadata and parameters in the DQ report.
}
