% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{valueCompletenessIndicator}
\alias{valueCompletenessIndicator}
\title{valueCompletenessIndicator}
\usage{
valueCompletenessIndicator(vm, vm_misg)
}
\description{
This function determines the Value Completeness Rate (dqi_cc_vcr), a generic indicator for assessing the completeness of mandatory data values. It also adds related metadata and parameters in the DQ report.
}
