% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{visualizeContradictions}
\alias{visualizeContradictions}
\title{visualizeContradictions}
\usage{
visualizeContradictions(ruleCol, valueCol, sumRow, vcPath, ...)
}
\description{
Function to visualize detected contradictions using a bar chart.
}
