% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{visualizeOutliers}
\alias{visualizeOutliers}
\title{visualizeOutliers}
\usage{
visualizeOutliers(itemCol, valueCol, sumRow, voPath, ...)
}
\description{
Function to visualize detected outliers using a bar chart.
}
