% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibRD.R
\name{checkCordDQ}
\alias{checkCordDQ}
\title{checkCordDQ}
\usage{
checkCordDQ(
  instID,
  reportYear,
  inpatientCases,
  refData1,
  refData2,
  dqInd,
  repCol,
  cl,
  itemCol,
  totalRow,
  oItem,
  ...
)
}
\description{
This function checks the quality of loaded data regarding selected DQ metrics for the CORD use case.
The default DQ dimensions are completeness, plausibility, uniqueness and concordance.
}
