% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cordDqLib.R
\name{checkUniqueOrphaCoding}
\alias{checkUniqueOrphaCoding}
\title{checkUniqueOrphaCoding}
\usage{
checkUniqueOrphaCoding(cl)
}
\description{
This function checks the uniqueness of RD cases coded with Orpha numbers
}
