% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCord.R
\docType{package}
\name{dqLib}
\alias{dqLib}
\alias{dqLib-package}
\title{dqLib: Tools for data quality assessment and reporting}
\description{
Contains methods for data quality analysis and reporting, especially in the field of Rare Diseases(RDs).
}
\author{
: Kais Tahar, University Medical Center Göttingen
}
\keyword{internal}
