% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\docType{package}
\name{dqLib}
\alias{dqLib}
\alias{dqLib-package}
\title{dqLib: Tools for data quality assessment and reporting on detected DQ issues}
\description{
This package contains methods for data quality analysis and reporting on identified quality issues, especially in clinical research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KaisTahar/dqLib}
}

}
\author{
\strong{Maintainer}: Kais Tahar \email{kais.tahar@med.uni-goettingen.de} (\href{https://orcid.org/0000-0001-9683-0575}{ORCID})

}
\keyword{internal}
