% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\docType{package}
\name{dqLib}
\alias{dqLib}
\alias{dqLib-package}
\title{dqLib: Data Quality Library (dqLib): An R Package for Explainable and Traceable Assessments of Clinical Data Quality}
\description{
This package offers a range of metrics for assessing data quality (DQ) across the dimensions of completeness, plausibility, uniqueness, and concordance. The implemented functions allow users to select appropriate indicators and generate customized DQ reports. The definitions of DQ dimensions and indicators are guided by the conceptual framework for harmonized DQ assessments first introduced under DOI: 10.1055/a-2006-1018. Additionally, dqLib provides methods for DQ analysis and reporting on DQ issues, especially in clinical care and research. The reports provide adequate information to explain the detected DQ issues and help users trace them back to their sources and underlying causes. The current version also enables the detection and visualization of plausibility issues based on predefined logical and mathematical rules. Further details on DQ metrics and exemplary DQ reports/visualizations are given in the readme.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KaisTahar/dqLib}
}

}
\author{
\strong{Maintainer}: Kais Tahar \email{kais.tahar@med.uni-goettingen.de} (\href{https://orcid.org/0000-0001-9683-0575}{ORCID})

}
\keyword{internal}
