% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{getLongReport}
\alias{getLongReport}
\title{getLongReport}
\usage{
getLongReport(metrics, sheetList, design, path)
}
\description{
This function creates detailed DQ reports based on user-defined specifications to explain the detected DQ issues and help users address them.
The reports include user-selected metrics and provide detailed information on detected outliers, contradictions, missing values, and missing items.
}
