% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCore.R
\name{getReport}
\alias{getReport}
\title{geReport}
\usage{
getReport(repMeta, sheetName, df, path)
}
\description{
Function to save DQ reports in Excel and CSV files. 
The CSV file includes only the user-selected metrics, while the Excel file provides multiple spreadsheets detailing the detected DQ issues, the user-selected metrics, the employed DQ rules, and the semantics of the DQ metrics.
}
