% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqLibCord.R
\name{getReport}
\alias{getReport}
\title{geReport}
\usage{
getReport(repMeta, cl, td, path)
}
\description{
This function generates data quality reports about detected quality issues, and user-selected DQ metrics such as indicators and parameters
}
